Remove home and update landing page to have previous home layout
