+++ src/js/components/landing-page/fake-room/fake-player.jsx
