+++ src/js/components/landing-page/fake-room/fake-room.jsx
