+++ src/js/components/landing-page/fake-room/fake-sidebar.jsx
