+++ src/js/components/landing-page/fake-room/fake-swipe-tabs.jsx
