+++ src/js/components/landing-page/fake-room/fake-text.jsx
